refactor: rename util package to internal
