+++ cmd/cli/app.go
@@ -7,8 +7,8 @@ import (
 	"github.com/sinlov/gh-conventional-kit/command/subcommand_markdown"
 	"github.com/sinlov/gh-conventional-kit/command/subcommand_template"
 	"github.com/sinlov/gh-conventional-kit/constant"
-	"github.com/sinlov/gh-conventional-kit/utils/pkgJson"
-	"github.com/sinlov/gh-conventional-kit/utils/urfave_cli"
+	"github.com/sinlov/gh-conventional-kit/internal/pkgJson"
+	"github.com/sinlov/gh-conventional-kit/internal/urfave_cli"
 	"github.com/urfave/cli/v2"
 	"runtime"
 	"time"
