+++ command/subcommand_new/subCmdNew.go
@@ -3,7 +3,7 @@ package subcommand_new
 import (
 	"github.com/bar-counter/slog"
 	"github.com/sinlov/gh-conventional-kit/command"
-	"github.com/sinlov/gh-conventional-kit/utils/urfave_cli"
+	"github.com/sinlov/gh-conventional-kit/internal/urfave_cli"
 	"github.com/urfave/cli/v2"
 )
