+++ command/subcommand_template/subCmdTemplate.go
@@ -9,9 +9,9 @@ import (
 	"github.com/sinlov/gh-conventional-kit/command"
 	"github.com/sinlov/gh-conventional-kit/command/common_subcommand"
 	"github.com/sinlov/gh-conventional-kit/constant"
+	"github.com/sinlov/gh-conventional-kit/internal/filepath_plus"
+	"github.com/sinlov/gh-conventional-kit/internal/urfave_cli"
 	"github.com/sinlov/gh-conventional-kit/resource/template_file"
-	"github.com/sinlov/gh-conventional-kit/utils/filepath_plus"
-	"github.com/sinlov/gh-conventional-kit/utils/urfave_cli"
 	"github.com/urfave/cli/v2"
 	"os"
 	"path/filepath"
